+++ netbox_documents/api/views.py
@@ -1,12 +1,17 @@
 from netbox.api.viewsets import NetBoxModelViewSet
 
 from .. import models, filtersets
-from .serializers import SiteDocumentSerializer, DeviceDocumentSerializer, DeviceTypeDocumentSerializer, CircuitDocumentSerializer 
+from .serializers import SiteDocumentSerializer, LocationDocumentSerializer, DeviceDocumentSerializer, DeviceTypeDocumentSerializer, CircuitDocumentSerializer 
 
 class SiteDocumentViewSet(NetBoxModelViewSet):
     queryset = models.SiteDocument.objects.prefetch_related('tags')
     serializer_class = SiteDocumentSerializer
     filterset_class = filtersets.SiteDocumentFilterSet
+
+class LocationDocumentViewSet(NetBoxModelViewSet):
+    queryset = models.LocationDocument.objects.prefetch_related('tags')
+    serializer_class = LocationDocumentSerializer
+    filterset_class = filtersets.LocationDocumentFilterSet
 
 class DeviceDocumentViewSet(NetBoxModelViewSet):
     queryset = models.DeviceDocument.objects.prefetch_related('tags')