+++ netbox_documents/models.py
@@ -19,6 +19,19 @@ class SiteDocTypeChoices(ChoiceSet):
         ('other', 'Other', 'gray'),
     ]
 
+class LocationDocTypeChoices(ChoiceSet):
+
+    key = 'DocTypeChoices.location'
+
+    CHOICES = [
+        ('diagram', 'Network Diagram', 'green'),
+        ('floorplan', 'Floor Plan', 'purple'),
+        ('purchaseorder', 'Purchase Order', 'orange'),
+        ('quote', 'Quote', 'indigo'),
+        ('wirelessmodel', 'Wireless Model (Ekahau)', 'yellow'),
+        ('other', 'Other', 'gray'),
+    ]
+
 class DeviceDocTypeChoices(ChoiceSet):
 
     key = 'DocTypeChoices.device'
@@ -149,6 +162,124 @@ class SiteDocument(NetBoxModel):
             raise ValidationError("A document must contain an uploaded file or an external URL.")
         if self.document and self.external_url:
             raise ValidationError("A document cannot contain both an uploaded file and an external URL.")
+
+    def delete(self, *args, **kwargs):
+
+        # Check if its a document or a URL
+        if self.external_url == '':
+
+            _name = self.document.name
+
+            # Delete file from disk
+            super().delete(*args, **kwargs)
+            self.document.delete(save=False)
+
+            # Restore the name of the document as it's re-used in the notifications later
+            self.document.name = _name
+        else:
+            # Straight delete of external URL
+            super().delete(*args, **kwargs)
+
+
+class LocationDocument(NetBoxModel):
+    name = models.CharField(
+        max_length=100,
+        blank=True,
+        help_text='(Optional) Specify a name to display for this document. If no name is specified, the filename or url will be used.'
+    )
+
+    document = models.FileField(
+        upload_to=file_upload,
+        blank=True
+    )
+
+    external_url = models.URLField(
+        blank=True
+    )
+
+    document_type = models.CharField(
+        max_length=30,
+        choices=LocationDocTypeChoices
+    )
+
+    site = models.ForeignKey(
+        to='dcim.Site',
+        on_delete=models.CASCADE,
+        related_name='document'
+    )
+
+    location = models.ForeignKey(
+        to='dcim.Location',
+        on_delete=models.CASCADE,
+        related_name='documents'
+    )
+
+    comments = models.TextField(
+        blank=True
+    )
+
+    class Meta:
+        ordering = ('-created', 'name')
+        verbose_name_plural = "Location Documments"
+        verbose_name = "Location Document"
+
+    def get_document_type_color(self):
+        return LocationDocTypeChoices.colors.get(self.document_type)
+
+    @property
+    def size(self):
+        """
+        Wrapper around `document.size` to suppress an OSError in case the file is inaccessible. Also opportunistically
+        catch other exceptions that we know other storage back-ends to throw.
+        """
+        expected_exceptions = [OSError]
+
+        try:
+            from botocore.exceptions import ClientError
+            expected_exceptions.append(ClientError)
+        except ImportError:
+            pass
+
+        try:
+            return self.document.size
+        except:
+            return None
+
+    @property
+    def filename(self):
+        if self.external_url:
+            return self.external_url
+        elif self.document:
+            filename = self.document.name.rsplit('/', 1)[-1]
+            return filename.split('_', 1)[1]
+
+    def __str__(self):
+        if self.name:
+            return self.name
+
+        elif self.external_url:
+            return self.external_url
+
+        elif self.document:
+            filename = self.document.name.rsplit('/', 1)[-1]
+            return filename.split('_', 1)[1]
+
+        else:
+            return ""
+
+    def get_absolute_url(self):
+        return reverse('plugins:netbox_documents:locationdocument', args=[self.pk])
+
+    def clean(self):
+        super().clean()
+
+        # Must have an uploaded document or an external URL. cannot have both
+        if not self.document and self.external_url == '':
+            raise ValidationError("A document must contain an uploaded file or an external URL.")
+        if self.document and self.external_url:
+            raise ValidationError("A document cannot contain both an uploaded file and an external URL.")
+        if self.location.site != self.site:
+            raise ValidationError("Location must belong to Site.")
 
     def delete(self, *args, **kwargs):